Fix cargo infer tests
Summary:
We do a check to make sure we're not type checking a whole project if it has a build system, since that might take down machines. This happens in getting configs, but the way infer works looks like it's trying to do a whole check.

To fix this, we expose a new function for just getting a config for a directory, which is the thing infer needs.

Reviewed By: stroxler

Differential Revision: D84165673

fbshipit-source-id: 5f6e72a664be926da1ed618f4c26a7c91e4f5614