Check for compatibility when inheriting from multiple classes for fields (#1196)
Summary:
This diff is an attempt to address the issue in https://github.com/facebook/pyrefly/issues/943.

When a class inherits from multiple base classes, we check whether the intersection of the field types is `never`. If so, we raise an error.

Next step: Handle method overrides / overloads.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1196

Reviewed By: stroxler

Differential Revision: D84155735

Pulled By: yangdanny97

fbshipit-source-id: 1b1e526951ae3ed79034d188fe2cb413d9027b26