The `orelse` merge of a loop is not a loop merge
Summary:
A "loop merge" is a merge that includes a loopback edge and has to deal with
things like recursive variables. The `orelse` merge of a loop with breaks
is not this - the Phi nodes are all distinct from the loopback merge (which
is a couple lines above).

This diff marks it as not a loop. Doesn't seem to produce any visible behavior
change which makes sense - treating it as a loop really only had the effect of
creating unnecessary `Default` bindings and since these phis aren't recursive
it typically won't matter whether we have a default.

Reviewed By: rchen152

Differential Revision: D84174338

fbshipit-source-id: 0c0af040b673a416763f72088a5955e8abfd514c