Represent `Loop` more clearly
Summary:
I'm already reasonably happy with flow-merging for loops (because we can eliminate
100% of narrows in the no-reassignment case we're already getting pretty close
to the best binding-time behavior possible).

But in order to improve flow merging for forks, I need the inner flow-merging
logic (which is shared) to be aware that the base flow is special. As a result,
I need to change the representation of `Loop` so that it's easier to explicitly
pass the base flow around instead of treating it like just another exit.

I think the resulting code is actually a little easier to understand anyway;
the base flow isn't really an exit, so just calling it "base" is honestly a bit
clearer.

Reviewed By: rchen152

Differential Revision: D84174332

fbshipit-source-id: 0c04f140c04f9be5018975c296c5d612805a9293