Fix upward search
Summary:
Minor bug from my file group upward find change a while ago, where a cached result won't be used if a closer result is available to the search start, even if it's in a lower priority group.

This diff fixes that to make sure a higher priority result is always used, even if it's cached.

Reviewed By: kinto0

Differential Revision: D84167152

fbshipit-source-id: 1f7e528ec48a58bb0c381e5a75406b5ecde1f1c8