Fix bug in override graph
Summary:
We currently have a bug in our override graph building.

The problem is in `get_super_class_member`: we use the module context of the derived class when looking up information from the base class. This leads to using bindings from the wrong module. In some rare cases, the binding keys match, and we get totally unrelated information.

This seems to be a general problem with our current design: we need to be careful every time we lookup information, we need to use the right bindings table for the class or function we are looking up.

I added cheap assertions in most of the functions where we make that assumptions, to catch problems in the future.

Reviewed By: tianhan0

Differential Revision: D84181520

fbshipit-source-id: a3f7d5e699c90eed6f4047487862ec6cebd7e451