Add inlay_hints for position function arguments.
Summary:
This diff adds inlay_hints for function arguments as long as the inlay_hint_config.call_argument_names option is not Off.

Previous tests are also updated to properly test this new feature.

There are 2 main ways that I thought to achieve this functionality:
1. Try to use the bindings already present in the function to get the call information that we need
2. Do another walk of the AST to get the call information and then generate hints.

After some research I realized there is limitation with trying to get it from the bindings:

The problem is that **not all calls have their own binding**. Only top-level expressions that need type inference get bindings. Nested calls within larger expressions won't have their own binding entry.

For example:

```
result = my_function(10, "hello", True)  # This call gets a binding
 x = my_function(10) + other_function(20)  # The addition might get a binding, but the two calls might not
```

So if you only iterate over bindings, **you'd need to recursively walk the Expr trees stored in each binding to find nested calls** - which is essentially still an AST walk, just a more limited one.

For this reason I think that it may make more sense to keep a full AST walk where we have access to more information. It may also make it easier in the future if we want to expand on this feature.

Reviewed By: kinto0

Differential Revision: D84090378

fbshipit-source-id: 4332343d073e5b7ac5ea2c2de28fb0a4f36e92cd