Rename `Contained` to `Partial`
Summary:
I renamed a lot of things a couple weeks ago to use the terminology
"partial type" and "partial type inference".

I didn't rename the variable type because at the time Sam was in the
middle of a big solver refactor and I didn't want to create unneeded
merge conflicts. Now seems like a good time to make the rename.

Reviewed By: rchen152

Differential Revision: D84238576

fbshipit-source-id: 40c99169172f0f843054453719f5358957b6146f