Push `base` down to the `merge_flow` API
Summary:
Re-wire the higher-level flow merging logic so that `base` is always
passed explicitly. The inner logic is unchanged (we ignore the base
in most cases, for loops we push it to the branches), but now all of the
plubming is ready for me to rework the `merge_flow` and `merge_flow_info`
logic so that we can actually use the base flow.

Note that in order to avoid an extra clone in the loopback merge of loops
that might not be needed if there are no breaks, I preemptively make an
optional cloned base ahead of time. This is slightly harder to read but
needed to avoid an extra clone on most loops (I'm pretty sure a majority
of loops in practice have no base, so this is potentially a lot of clones
we are avoiding).

Reviewed By: rchen152

Differential Revision: D84232166

fbshipit-source-id: 7fa7b69c96a8bcc2a33edcf785019fe0dd806da7