Rename some things
Summary:
It was never really obvious why to name one set of flow `flows` vs another
one `branches`, I just like having two different names and avoiding shadowing.

But I'm about to switch the interface so that the `base` flow gets passed
in expclicity so that we can use it for better Phi behavior, and at that point
it seems pretty clear that it's better to name the parameters `base` and
`branches` rather than `base` and `flows`.

So this diff just flips the existing variable names around.

Reviewed By: rchen152

Differential Revision: D84232165

fbshipit-source-id: a15742dae9871e1af8438576229a44d20196e69f