Further elaborate `MergeStyle`, simplify the stack of merging functions
Summary:
If we distinguish exclusive vs inclusive branching merges, we can eliminate
`merge_into_current` and just always set the current flow to the result of
merging.

This is slightly simpler as-is since we can get rid of some boolean flags and
an intermediate layer in the call stack. It also makes it much easier for me to
push down the handling of base flow (which we want to start using even for
exclusive and bool op merges where it isn't directly part of the merge, because
that will allow us to handle Phis much better).

Reviewed By: rchen152

Differential Revision: D84232169

fbshipit-source-id: bfa7f6257b74cdd3b83957c7817a860801b95e14