Simplify the implicit else/wildcard handling in `finish_fork_impl`
Summary:
We had some complicated swapping logic in place to avoid cloning base here, but
I'm about to implement a change that requires the base flow to improve Phi behavior
anyway.

As a result, it's nice to use much simpler - if slightly less efficient - logic
first so that the upcoming diffs are easier to read.

Reviewed By: rchen152

Differential Revision: D84232170

fbshipit-source-id: 4ca9d67545d28b664bc583d1bc70411fe2c2f26f