Rename some vars
Summary:
The var formerly known as `downstream_idx` is just the merged idx.

I named it downstream and contrasted with upstream when I was doing the phi
elimination for loops because that was a reasonable way to think about the
optimization, but at this point those names aren't helping me anymore.

Reviewed By: yangdanny97

Differential Revision: D84256249

fbshipit-source-id: 3cc0933b79a0222da759b4b03c570a35dfd31fd1