Re-introduce MergeItem to track base FlowInfo
Summary:
This diff re-introduces the `MergeItem` type (which used to contain different
data, I eliminated it when I made merging powered directly by `FlowInfo`
because we didn't need really it) so that we can track both the base flow and
branches as we're building up merge data for each name in the flow merge.

For now, the only thing we actually do with this is push the base flow for
loops (which we were already doing before, but at the whole-Flow leve in `merge_flow`,
now we can do it per name). But with this change we now have the ability to actually
see in `merge_flow_info` whether or not there was a base flow info, so we are continuing
or work toward having everything put together in a way that will let us make improvements
to the flow merging logic.

There should be no behavior change. At the moment there's a lot of fiddly
counting-related logic where we have to be pretty careful about loops in order
to keep uninitialized local checks working; this is probably unavoidable if we want
to share most of the logic between forks and loops (which is probably worth the cost).

Reviewed By: rchen152

Differential Revision: D84256250

fbshipit-source-id: 60aa5d70a7960d521d484608aca09dd93b22c365