Rework the default handling
Summary:
It's a little weird to pull the default value from the first flow info,
because that potentially isn't even coming from the base flow.

Let's get it from the base instead. This technically introduces a behavior change
in the bindings layer because now a name defined in a loop won't get a default, but
that's actually not relevant to loop handling because we can't actually read from
the loop phi anyway (since it wasn't in the base flow). If anything in the loop
tried to read the uninitialized variable, it would just get the anywhere-style
static scope lookup anyway, so nothing is really changing.

Any behavior changes are small enough that I'm not even sure how to construct
examples where it would matter, although it's possible such examples exist
in cyclic cases where Pyrefly doesn't generally work anyway:
- In a normal loop, we'll create a Phi with no default (so that forcing it
  is Unknown) for a variable defined in the loop. This probably has no effect
  since there's no recursive looping behavior, but it could potentially show
  up in the result of breaking a cycle from recursive return inference or
  some other non-loop cycle.
- In nested loops, we'll now set the default of a name that's defined by an
  inner merge to the merge itself; this is reasonable, since the merge isn't
  going to be recursive. It probably has no effect, but again it's possible
  there are cyclic cases where it might matter.

At any rate, I'm certain that any changed behaviors were never really intended
anyway (setting a default for a name not in base doesn't make sense), and
zero unit tests change.

Reviewed By: rchen152

Differential Revision: D84256252

fbshipit-source-id: 87af5696b8632412fb3489a2a0edfdcdbb45a3fb