Simplify special_case_enum_attr_lookup, fixing a bug
Summary:
I found this function a little hard to follow because (1) it handled `value` and `_value_` together, and (2) it did more than it needed to because it sometimes called get_instance_attribute when it could just return None.

I restructured it as follows:
1. Check if we have an attribute that needs special handling, returning None if not.
1. If the attribute is _value_, infer a type for it.
1. If the attribute is value, look up _value_ and turn it into a property.

In addition to being more readable, this new version also fixes a bug that was in the old one.

Reviewed By: stroxler

Differential Revision: D84324317

fbshipit-source-id: ddcb73009836b04fce2a401afe8e16847a6f27c8