tighten up `__exit__` check for context manager
Summary:
Joren says you can either call it with `type[BaseException], BaseExeption, TracebackType` or `None, None, None`, but not a mix.

Our current check gives a false positive for users who define this beahvior with 2 overloads.

To fix this, we'll just call the function twice with the different argument sets and union the results.

To reduce noise when we emit errors, we forward the error directly if there's only 1 error during the call, but consolidate the errors if there's >1 error.

fixes https://github.com/facebook/pyrefly/issues/1221

Reviewed By: stroxler

Differential Revision: D84353112

fbshipit-source-id: 124efdc169e839680a546ab24fb875b8576b4fdc