Add docstrings to attribute.
Summary:
This diff updates attributes to have a docstring. In a subsequent diff we will add these docstrings to attribute completion.

The initial approach that was taken was to walk the entire AST for each method and find the covering nodes of each function definition. From there we would be able to find the function definition and get the docstring from that.

Instead of that approach, we now add the docstring_range onto ClassFieldProperties. This approach allows us to avoid an extra AST walk. Then in resolve_attribute_definition we can grab the docstring range from the attribute and add it to our completion items.

Reviewed By: kinto0

Differential Revision: D84271890

fbshipit-source-id: d989b186b246aeeb71cff92a9fe6376fbcc10163