Change expected test result for handling empty case
Summary:
I tested the example case at runtime and the revealed type was `<class 'django.utils.functional.lazy.<locals>.__proxy__'>`

I do have a diff that can widen the type, but I don't think we should do that? the type of the field looks correct to me in that it matches runtime so we should probably diverge from mypy here.

Reviewed By: ndmitchell

Differential Revision: D84380516

fbshipit-source-id: e5c6646be97ebdc3dd21348069fb473886b102ae