Create a LoopPhi binding
Summary:
This is a BE change I've wanted to make ever since removing all use of
Default except for the recursive phi: we should just use a special binding
for loop phi instead of boxing a binding in a Default binding that is much
more general than we need.

This will eventually help if we make changes to how loops are modeled, but
it also makes it easier to add logic to Phi that doesn't make sense for
LoopPhi, since now they are decoupled.

That's my real motivation: I want to add an enum that potentially includes
the base idx and whether or not we're looking at just narrows of it to
Phi, and not all of the enum behaviors will make sense for LoopPhi so
now seems like a good time to make the BE change I've been wanting to make
for a while.

Reviewed By: grievejia

Differential Revision: D84381062

fbshipit-source-id: 2898b747d376f9898516522da9cafa3a23eafc43