Move `JoinStyle` to `type_info`, make it generic
Summary:
To actually customize join behavior cleanly, we need to do it per-facet.

For this, a perfect setup is to mirror the binding and type structure by
making the join style generic so that we can turn a `JoinStyle<Idx<Key>>`
into a `JoinStyle<TypeInfo>` by looking up the key, and later turn that
into a `JoinStyle<Type>` as we pick a particular spot on the `TypeInfo`
tree.

Reviewed By: yangdanny97

Differential Revision: D84381059

fbshipit-source-id: 3b426082739a0734daea7929b095ab27c016bdfe