Add monitoring for slow functions
Summary:
# Problem
We found out about a bottleneck when exporting definitions for a specific module. This was hard to debug.

# Solution
This diff adds a slow function monitoring tool, which warns when a function takes more than a certain of time to complete. It warns when it reaches a certain duration threshold, and then warns about at 2*threshold, 4*threshold, etc.
This is implemented efficiently and safely with a single monitoring thread, rather than using one thread per monitored function, or using (unsafe) UNIX signals.

Reviewed By: tianhan0

Differential Revision: D84498340

fbshipit-source-id: dc7fa705a658e949a8d5f1129ab68977591e649b