Add a test where Pyrefly misbehaves in flow-merging for gradual code
Summary:
Pyrefly currently unions types blindly in flow merging. This behavior is reasonable in a
static type system that is high-fidelity on merges, but Pyrefly is neither:
- We have the `Any` type, which is pretty special and typically expresses logic
  that is beyond the static type system's understanding (e.g. duck typing,
  control-flow-dependent types that go beyond what narrowing can track, etc)
- We lack intersection and negation types, which are required in order for
  the set operations described by narrow-and-merge to always produce correct answers

This diff adds tests tracking how the lack of intersections and over-eager use of
static types affects gradual code: here both `x` and `y` are `Any`, the user has
explicitly told us that their types are beyond our understanding.

It's reasonable for us to narrow in simple control flow where we can see a type,
but merging those types back into a union with `Any` is catastrophic: in an
actual gradual codebase with big functions, this could light up almost every
line below the first nontrivial operation with a type error!

This has been my single highest-priority issue with flow merging, we cannot do
a beta launch with behavior this severe.

Reviewed By: yangdanny97

Differential Revision: D84381066

fbshipit-source-id: 60431e6d92de615e6512174ae7bef6f713249009