Add a `join_types` helper
Summary:
I'm almost ready to optimize join behavior based on the JoinStyle,
which at this point we've passed from bindings to answers solving
and then pushed down the TypeInfo tree.

We need one more layer before we can start implementing the type-level
logic, which is to thread it into an actual type-level join.

This diff introduces a `join_types` helper function so that we have
a place to add additional join logic on top of the basic `union_types`
unioning logic.

Reviewed By: yangdanny97

Differential Revision: D84381063

fbshipit-source-id: 1e74d1bd5404f5bb0969309f36fc1adb71ab6f99