Add unannotated local variables to symbol table in VScode
Summary: This diff will add unannotated local variables to VScode. To do this, we add a new arm to the match statement in recurse_stmt_adding_symbols within the symbols function. If we have a valid target, we pull the name from it and create a new document symbol struct which is then pushed into the symbols collection.

Reviewed By: kinto0

Differential Revision: D84535534

fbshipit-source-id: 892065486cf43448962e83c4186762b3b7476e31