Add annotated local variables to symbol table in VScode
Summary:
This diff will add unannotated local variables to VScode. To do this, we add a new arm to the match statement in recurse_stmt_adding_symbols within the symbols function. If we have a valid target, we pull the name from it and create a new document symbol struct which is then pushed into the symbols collection.

For annotated assignments we will also grab the type and stick it in the detail field with the following code block:

```
                            detail: Some(
                                module_info
                                    .code_at(stmt_ann_assign.annotation.range())
                                    .to_owned(),
                            )
```

Reviewed By: kinto0

Differential Revision: D84535533

fbshipit-source-id: ad3bd86d8c75140b42230b6ac48aa637f2dbf658