Add failing test for class confusion in IDE
Summary:
In the IDE, we check an "in memory" version of open files. If that file exists in a cycle, the in-memory version of a class can be compared to the on-disk version.

In this test, we open `foo.py` and check it. We call `Bar.bar(self)`, where `self` is the in-memory `Self@Foo`.

`Bar.bar` is annotated as taking `Foo`, but this resolves to the on-disk class.

Reviewed By: kinto0

Differential Revision: D84540242

fbshipit-source-id: b8ba2c28a1105b03d5e99e73adbe6f4153387979