Add support for TSP snapshot request (#1194)
Summary:
This PR adds another request handler for TSP = "typeServer/getSnapshot". The snapshot request is how the client side of a TSP connection indicates which 'state of the world' it's currently querying about. The server side keeps a 'snapshot' to indicate when the state of the world has changed.

Internally I used the 'epoch' and the 'RecheckFinished' event to indicate what the snapshot is and when it's changed.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1194

Reviewed By: stroxler

Differential Revision: D84510914

Pulled By: kinto0

fbshipit-source-id: f645855790414b7cb4d7525bf2912019ccbfac4b