Verify that issue 1246 is fixed
Summary:
For issue 1246, I had initially planned to prune narrows so that
any type that isn't a subtype of the base gets discarded in the union.

This would be correct, since "true narrow" types are always intersections
so any such narrow is always a lossy type that is potentially going to lead
to a bad join with lots of false positives downstream, but is potentially a bit
expensive.

I actually realized that the previous fix to simplify types is enough to handle
all practical examples of this, because in practice the bad cases with awful
merges involve merging a negated base flow, and when we do that in the presence
of a lossy narrow type the original type always comes through.

This diff adds two more test cases - one of issue #1246 as reported in a boolean
operation and another simpler one with an `if` - verifying that the behavior
is pretty decent now.

I think we can skip the pruning - since we eventually need intersection types
of some sort anyway to fix some practical examples users are hitting (I can't
seem to search our issues anymore, there are too many, but I know someone had
a protocol example in real-world code that Pyrefly can't handle), it seems okay
to defer further improvements to flow merging until then because intersection
types will solve the problem once and for all.

Fixes https://github.com/facebook/pyrefly/issues/1246

Reviewed By: rchen152

Differential Revision: D84381061

fbshipit-source-id: cd50d23c0a8f496ccb9eecbacde62d8e84c468ab