Make flow-merging-joins respect explicit and implicit `Any`
Summary:
This diff fixes the biggest single concern I had with Pyrefly flow merging, which
is that taking an `Any` and permanently unioning it with concrete types because
of branch-specific logic can be catastrophic in gradual codebases: the `Any` type
often means "this variable's type isn't statically analyzable", and that can
mean things like duck typing or control-flow-dependent types that go beyond what
static checkers understand. Forming a union and complaining on all downstream
uses is pretty clearly not right, and could easily cause a single narrow to
turn almost an entire large function red even when every variable is typed as `Any`.

Interestingly we actually have no open issues on this (although I do remember
a bug report about it - maybe just verbally at PyCon?) but I am firmly convinced
this fix had to go in before a beta release.

With this diff, we still allow the merge in the case where all branches of the
merge have non-`Any` types. This is somewhat debatable, especially for explicit
`Any`, but I think there's often a good chance the user actually intended us to
infer a concrete type; it's easy to change the behavior later if we see a need.

I'm only 100% sure we need to do this for explicit `Any`. I opted to do this
for both explicit and implicit `Any` because if we want Pyrefly to be adoptable
on gradually typed code, implicit `Any` will be common and preserving
narrows across control flow merging is a really big potential source of false
positives. This is somewhat debatable and easy to change.

I opted *not* to do this for `Any` coming from errors, which I also feel
confident about - those `Any`s aren't really gradual types, they are escape
hatches to reduce downstream noise after an upstream type error, and I think
applying normal flow merging to them is probably a good thing, especially
for the UX around *fixing* the upstream error and what kinds of new downstream
errors appear.

Reviewed By: rchen152

Differential Revision: D84381072

fbshipit-source-id: 27d2242629d693351ae2e81c0273d322935de9fd