Simplify narrow-only joins when base is a subtype of the result
Summary:
We cannot currently simplify unions efficiently because without some kind of subtype
check caching it is quadratic to do so.

But in the special case of a union of narrows, we already know the base type so it's
linear to check for simplification. The resulting type will be vastly more performant
downstream, plus it eliminates a verbose and potentially confusing unnecessary
union in user-facing hover results, error messages, etc.

NOTE: I avoid trying to simplify reassignments because things are a bit more subtle
there - for example one branch might assign to `Any`. I think we should just be
patient and wait for proper union simplification (which will probably happen when
we make a type heap) rather than getting into this; for narrow-only joins these
problems don't occur since "true narrow types" are always intersections, so
`joined_ty <: base_ty` would be a tautology if our type system were high enough
fidelity.

Reviewed By: rchen152

Differential Revision: D84381060

fbshipit-source-id: ae11ec5a5b99200d5818ce916444de25c737a4c9