Add pydantic types module name
Summary:
We want to strictly typecheck `StrintInt` and similar types, which have the form `StrictInt = Annotated[int, Strict()]`

One approach can be that we examine this stub and navigate the structure of the type for the strict marker, which will enable us to mark those types as strict.

The issue is that by the time we reach the check where we extract the rootmodel type and convert it to `Any`, we have already stripped the strict marker from the type.

I think the solution would be to mark the strictness earlier in the process by matching on `pydantic.types` and then examining the structure of the type Annotated[T, Strict()]` and marking it as strict in the presence of the marker. It's unclear to me yet how early in the pipeline we have to go to to this, since it seems to be happening before even the metadata layer.

The other (easier option) is to match on the name of the type, but that seems like the less robust solution. So my plan is to look into the first option.

Reviewed By: kinto0

Differential Revision: D84654810

fbshipit-source-id: 4bdbce48f82826e50ce8db578650b3e8ce01fbba