Add synthesized fields in class definitions
Summary:
We were not including synthesized fields in classes because `Class::fields()` only retuns "regular" fields. We need to look up the `KeySynthesizedClassFields` key in the answer table. This diff fixes the problem.

Note that we don't want to export class fields that have a function-like type, such as `__init__`, etc.
In a follow-up, we will actually export those in "function_definitions".

Reviewed By: tianhan0

Differential Revision: D84606379

fbshipit-source-id: 5f3ce1d7c361aefa0679227d1d102b6b4917c133