rename test to make it more clear
Summary: this test takes a second to read / understand. language clients will sort the responses by sorttext, so it's not necessary for us to sort it by sorttext ourself. but that's what this test is testing for, and it might be necessary for assumptions about other tests, so i'm renaming it.

Reviewed By: tpolasek, ndmitchell

Differential Revision: D84712429

fbshipit-source-id: ed94007bd791d03d8de7cfb602de9dc0b5d3bb91