Warn on attempt to type-check all of ~
Summary:
It's sadly easy to accidentally run pyrefly over all of `~`, which can cause it to seemingly hang.

I originally wanted to make pyrefly emit a warning if it detects that it's being run on a very large number of files, but I couldn't find a good way to do that, since it doesn't even finish listing all the files.

Instead, I added a warning that fires on `pyrefly check` being run with a default project-includes in the `~` directory or one of its ancestors. This also produces a warning when `pyrefly init` is run in `~` or an ancestor. Pyrefly will not warn if you specify files to check on the command line, or (implicitly or explicitly) use a config with non-default project-includes, or explicitly write `pyrefly check ~`.

Reviewed By: stroxler

Differential Revision: D84667527

fbshipit-source-id: 2a9a233d5d4d56d79677945ebffeef2cba632d50