Use `expand_vars` for var expansion
Summary:
The term "expand" is too vague for a codebase this size - we already have
overloaded the term "expand type" to mean expansion in the sense of the
overload specification, which has exactly nothing to do with var expansion.

In addition, because calls are spread out and use three different functions,
the ability to easily search for calls to any of the functions is handy, and
that's easier if they have a recognizable name; `expand` isn't sufficiently
descriptive to be a good search (lots of false hits) but `expand_vars` is.

Reviewed By: rchen152

Differential Revision: D84829659

fbshipit-source-id: b8f38c65d9ec666d54f7923dc2ca18fc859f143d