Rename `FilesConfigConfigurer` to `DefaultConfigConfigurerWithOverrides`
Summary:
rename-yes

There was `DefaultConfigConfigurer` and there was `FilesConfigConfigurer`. It appears that the latter has very little to do with "files" -- the only difference between the latter and the former is that the latter takes CLI config override flags into account, but the former does not.

This diff moves `FilesConfigConfigurer` and have it sit next to `DefaultConfigConfigurer`. I renamed `FilesConfigConfigurer` to `DefaultConfigConfigurerWithOverrides` to make the differences between the two more clear. I also renamed the wrapper funcs around these two `default_config_finder()` and `default_config_finder_with_overrides`, so client code now don't have to worry about the notion of "config configurer".

Reviewed By: vladima

Differential Revision: D84774621

fbshipit-source-id: 920a635b7e1f47a37a057183720a826fd0db589a