Add annotated metadata to TypeAlias
Summary:
I think we should go with storing the entire vector of types, because `strict()` can exist in the vector, along with other elements, such as `gt` and `lt`. So storing the whole vector would allow us to reason about more than just the strict keyword for  `Annotated`

RFC:
This is making certain structures too large. Is this acceptable or are there alternative thoughts on this?

Reviewed By: rchen152

Differential Revision: D84753228

fbshipit-source-id: 60e87f89bf27777c034c410fe55349a4dfc1a9e3