Intern PathBuf instead of ModulePathDetails
Summary:
Interning helps with performance because hashing/equality for PathBuf can be expensive. Pushing the intern in makes it easier to use just the path as an equality key for classes (next diff).

This change was considered and rejected in D84636337 (per ndmitchell, "simpler"). While I agree this is more complicated, I think the compexity is warranted. In fact, I think interned paths probably belong a bit higher in the stack, but I'm going to leave that for future work.

Reviewed By: stroxler, ndmitchell

Differential Revision: D84834587

fbshipit-source-id: 1b5f010fa3f5f941eb4dca1f34c0e119fbf31867