Add a `severity` field to the output JSON
Summary:
Our IG partners are asking for an access to the severity of each type errors Pyrefly emits. The info is currently hidden by Pyrefly with `--output-format=json` because Pyrefly is still using the legacy Pyre1 JSON format which does not include the severity field.

This diff attempts to add the extra severity field to the JSON format. This should be still compatible with Pyre1 since we are not removing / changing any field. The severity field can currently be one of 4 strings: `error`, `warn`, `info`, and `ignore`.

Reviewed By: tpolasek, kinto0

Differential Revision: D84861461

fbshipit-source-id: a50b14b45198dcf14f3f3fe84aa794ab680cd019