Extend ClassBases with strict flag
Summary:
We store  strictness info in `base_class_expr_untype` and we plan to return that into (we will change the return type to be a tuple), but first, we should extend the structure of ClassBases so we can propagate the info.

Another alternative could have been to keep returning tuples. I thought extending the structure would be cleaner but happy to hear further thoughts about that.

Reviewed By: rchen152

Differential Revision: D84938994

fbshipit-source-id: 85619a39c477ad2fac2d9692ff53cfdc8c932f12