Typeshed refactor: Extract find function out to re-usable function
Summary:
This diff extracts the find functionality out into its own function which can then be re-used later.

Overall the functions behave identically. The only difference is that instead of referencing self, we now pass in a BundledTypeshed struct to the function which is used.

In the future, this type can be expanded to be any other type of stub one they are implemented.

Reviewed By: kinto0

Differential Revision: D85011599

fbshipit-source-id: 8df8802678c664d4674e9cc20abdb53784637057