Resolve property setters and getters
Summary:
The main change is to add `assignment_targets` in the AST visitor, when
visiting an expression. This is used to decide whether a property getter
(determined by the go-to-definitions, which does not distinguish getters from
setters) is actually a property setter, when it is on the lhs of an assignment.

Reviewed By: arthaud

Differential Revision: D84848888

fbshipit-source-id: e407867ad7954260cfcaf7a76569c95da1872ff6