Add option for single task queue in LSP
Summary:
Right now, Pyrefly kicks off a bunch of build system rebuilds for each file opened, which can cause a bunch of extra sourcedb rebuilds to happen if files in different targets are opened one after another. To fix this, we should do something similar to our other build system language server, and only process the last build system task if there  are more than one pending.

There are a few different approaches we can take for this, but my thoughts right now are to give the `HeavyTaskQueue` the ability to consume as many tasks as are available, and only actually use the last task produced.

An alternative implementation would be to start giving queued tasks cancellation handles, but that might be overkill for now, since it would also require book keeping somewhere of any previously queued build system tasks, but would be more flexible for future heavy task operations we want to do.

Reviewed By: kinto0

Differential Revision: D84942840

fbshipit-source-id: 3bf86eee75835160fcd92a8c589ba5b1e9a1918d