Fix Reset issue not creating proper files for non code snippets.
Summary:
The default sandbox code example expects there to be a util.py file and a formatter function to be imported from it.

If you are in a non-default code example that does not have a util.py file and try to reset, the file will not be created and there will be an import error in the code.

This diff creates a new case where for non code snippets, we ensure that the util.py file is created on reset.

Reviewed By: yangdanny97

Differential Revision: D85151323

fbshipit-source-id: 1a3fed0819812336ad96558ac954b556a86e28c8