Change how default is fetched
Summary:
I don't think the existing default logic, which is only using the default
if the Calculation has already been created, makes sense.

Maybe there's some notion of performance optimization that went into this, but
this is transparently nondeterministic, it's very sensitive to the exact order in which
bindings are evaluated. It's almost always a bug to write code like this in Pyrefly.

Let's use `get_idx` instead, this is much more likely to behave deterministically,
and it's also a change I need to make if I want to meaningfully restructure how
loop recursion works.

Reviewed By: yangdanny97

Differential Revision: D84724940

fbshipit-source-id: e1f54b33e3f4da40b5aa301fb74dd9bbabb3fbbe