In subset.rs, use a single match to describe the semantics of cases
Summary:
To actually understand the current code in `is_subset_eq_var` you have to
follow three layers of matching on the same `Variable`:
 - An outer match in the big match statement
 - Inside of that, hooks to modify a block's behavior for `Quantified`
 - And separately, another match with divergent behavior in `maybe_promote`

I'd like to avoid this - understanding the semantics here is pretty important,
and it's possible to use exactly one match for all three; in the process of
doing that we can also avoid wildcards, so that you can see what the behavior
is locally without having to jump up and reference the `Variable` enum
(at the moment, you have to do set subtraction to answer a question like "which
types of variables *don't* use promoted types).

This diff makes that change. It costs us a tiny bit of duplication in the outer
match, but allows us to avoid wildcards and eliminate the `maybe_promote` helper.
In the process some of the existing comments lose a logical place to be, so I
explain promotion in the doc comment at top.

Reviewed By: rchen152

Differential Revision: D85076869

fbshipit-source-id: bca49c94eb81409e70eae3b85b4b9b741fed49ae