Update comment to document that relative path for typeshed is typeshed/stdlib.
Summary:
This diff simply updates a comment in module_path.rs to say that the relative path for typeshed is typeshed/stdib.

We saw this comment initially and were concerned that this may cause issues as we add other types of stubs besides the stdlib stubs that we already bundle. If both stdlib and other types of stubs share a relative root path, we may potentially materialize them together and be using stubs from a module that we do not intend to.

I did some digging to see how we could change this and found this this comment may be outdated. Upon more investigation it looks like the relative path is typeshed/stdlib and not typeshed like we thought previously. How I can to this conclusion is included in the test plan section.

Reviewed By: kinto0

Differential Revision: D85038959

fbshipit-source-id: 4f40b6ab98e6bcbbb88f505ada28f29960e596d8