Add natural inference tests
Summary:
I wanted a place to add some tests, and realized that it would be useful to have a place specifically for "natural inference" tests.

Notably, I wanted to think a bit more about when/how to perform literal promotion, whether we want to preserve literal type information when instantiating covariant typevars, validate our behavior with tuples, etc.

The behavior isn't set in stone, so this is mostly documenting the consensus behaviors. That said, feel free to disagree with any specified behavior here.

Reviewed By: rchen152

Differential Revision: D85195835

fbshipit-source-id: ac9689c7f3d36df8a6d7cf741de4c5c795752e70