Add failing test for regressed literal promotion of tuples
Summary:
When the tuple literal is passed directly to a generic function call, we end up failing to perform literal promotion when instantiating.

The issue appears to involve contextual typing, which changed in D84866985. I confirmed that this regression happened in that diff by locally reverting.

Note that the following passes

```
def f[T](x: T) -> list[T]: ...
tup = (1,)
assert_type(f(tup), list[tuple[int]])
```

so clearly it's important that we are contextually typing the tuple literal expression against `T`. In general, a Var without a solution can not provide a hint, so maybe there is a more general issue here.

Reviewed By: rchen152

Differential Revision: D85203217

fbshipit-source-id: f8d1f1fede34b15f619b35e2f1a5cf7224fd943b